requirements.txt
# YOLOv5 requirements
torch>=1.7.0
torchvision>=0.8.1
torchaudio>=0.7.0
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
matplotlib>=3.2.2

# API requirements
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
python-multipart>=0.0.5
